backend/pyproject.toml
[tool.poetry]
name = "document-intelligence-backend"
version = "0.1.0"
description = "Async document processing backend"
authors = ["Kolade"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110"
uvicorn = "^0.29"
pydantic-settings = "^2.2"
python-json-logger = "^4.0.0"
sqlalchemy = "^2.0.45"
asyncpg = "^0.29.0"
alembic = "^1.17.2"
psycopg2-binary = "2.9.11"
python-jose = "^3.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.21"
minio = "^7.2.20"
boto3 = "^1.42.16"
redis = "<6.0.0"
celery = {version = "5.3.4", extras = ["redis"]}
pytesseract = "^0.3.13"
pillow = "^10.4.0"
pdf2image = "^1.17.0"
python-docx = "^1.2.0"
pdfminer-six = "^20251228"
pandas = "^2.3.3"
openpyxl = "^3.1.5"
google-generativeai = "*"
google-genai = "^0.4.0"
pgvector = "^0.3.6"
llama-index = "^0.12.0"
llama-index-vector-stores-postgres = "*"
llama-index-llms-gemini = "*"
llama-index-llms-ollama = "*"
llama-index-embeddings-gemini = "*"
llama-index-embeddings-ollama = "*"
slowapi = "^0.1.9"
asgiref = "^3.11.0"
pydantic = {extras = ["email"], version = "^2.12.5"}
pytest = ">=8.0.0"
aiosqlite = "^0.22.1"
python-magic = { version = "^0.4.27", markers = "sys_platform != 'win32'" }
python-magic-bin = { version = "^0.4.14", markers = "sys_platform == 'win32'" }
pypdf = "^5.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1"
pytest-asyncio = "^0.23.0"
anyio = "^4.0.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"